+++ R/spada.R
@@ -13,7 +13,7 @@
 #'
 #' @importFrom bsicons bs_icon
 #'
-#' @importFrom bslib accordion accordion_panel bs_theme card card_body
+#' @importFrom bslib accordion accordion_panel bs_add_rules bs_theme card card_body
 #'             card_footer card_header layout_column_wrap layout_columns
 #'             layout_sidebar nav_item nav_menu nav_panel nav_select nav_spacer
 #'             navset_card_pill page_navbar popover sidebar tooltip value_box
@@ -36,6 +36,8 @@
 #'             render_gt sub_missing sub_values tab_options
 #'
 #' @importFrom haven as_factor is.labelled read_sav write_sav
+#'
+#' @importFrom sass as_sass
 #'
 #' @importFrom shinyWidgets colorPickr updateColorPickr show_toast dropdownButton
 #'             radioGroupButtons statiCard